[FIX:sparkles:] Update to flask-ligand v0.4.0
The new version of 'flask-ligand' changes how the 'create_app' entry-point
works which required an update to how this example is initialized. I also
updated the documentation with some small tweaks and fixes for typos.

I also added new PyCharm run configurations that make it more convenient to
to launch a local Flask server and directly interact with the SwaggerUI docs
from within PyCharm only.